move `time_out_assert()` family back to tests
