+++ mypy-exclusions.txt
@@ -14,7 +14,6 @@ chia.rpc.util
 chia.rpc.wallet_rpc_client
 chia.simulator.full_node_simulator
 chia.simulator.keyring
-chia.simulator.time_out_assert
 chia.simulator.wallet_tools
 chia.ssl.create_ssl
 chia.timelord.timelord_api
@@ -106,6 +105,7 @@ tests.util.generator_tools_testing
 tests.util.test_full_block_utils
 tests.util.test_misc
 tests.util.test_network
+tests.util.time_out_assert
 tests.wallet.cat_wallet.test_cat_lifecycle
 tests.wallet.cat_wallet.test_cat_wallet
 tests.wallet.cat_wallet.test_offer_lifecycle
