+++ tests/core/full_node/full_sync/test_full_sync.py
@@ -13,15 +13,14 @@ from chia.full_node.full_node import FullNode
 from chia.full_node.full_node_api import FullNodeAPI
 from chia.protocols import full_node_protocol
 from chia.protocols.shared_protocol import Capability
-from chia.simulator.time_out_assert import time_out_assert
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.blockchain_format.sub_epoch_summary import SubEpochSummary
 from chia.types.full_block import FullBlock
 from chia.types.peer_info import PeerInfo
 from chia.util.hash import std_hash
 from chia.util.ints import uint16
-from tests.conftest import ConsensusMode
 from tests.core.node_height import node_height_between, node_height_exactly
+from tests.util.time_out_assert import time_out_assert
 
 log = logging.getLogger(__name__)
