+++ tests/core/full_node/test_transactions.py
@@ -10,11 +10,11 @@ from chia.consensus.block_record import BlockRecord
 from chia.consensus.block_rewards import calculate_base_farmer_reward, calculate_pool_reward
 from chia.full_node.full_node_api import FullNodeAPI
 from chia.simulator.simulator_protocol import FarmNewBlockProtocol
-from chia.simulator.time_out_assert import time_out_assert
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.peer_info import PeerInfo
 from chia.util.ints import uint32
 from chia.wallet.util.tx_config import DEFAULT_TX_CONFIG
+from tests.util.time_out_assert import time_out_assert
 
 
 class TestTransactions:
