+++ tests/core/mempool/test_mempool_fee_protocol.py
@@ -12,10 +12,10 @@ from chia.protocols.wallet_protocol import RespondFeeEstimates
 from chia.server.server import ChiaServer
 from chia.simulator.block_tools import BlockTools
 from chia.simulator.full_node_simulator import FullNodeSimulator
-from chia.simulator.time_out_assert import time_out_assert
 from chia.util.ints import uint64
 from chia.wallet.wallet import Wallet
 from tests.core.node_height import node_height_at_least
+from tests.util.time_out_assert import time_out_assert
 
 
 @pytest.mark.anyio
