+++ tests/core/server/test_dos.py
@@ -14,10 +14,10 @@ from chia.protocols.shared_protocol import Handshake
 from chia.server.outbound_message import Message, make_msg
 from chia.server.rate_limits import RateLimiter
 from chia.server.ws_connection import WSChiaConnection
-from chia.simulator.time_out_assert import time_out_assert
 from chia.types.peer_info import PeerInfo
 from chia.util.errors import Err
-from chia.util.ints import uint16, uint64
+from chia.util.ints import uint64
+from tests.util.time_out_assert import time_out_assert
 
 log = logging.getLogger(__name__)
