+++ tests/farmer_harvester/test_farmer_harvester.py
@@ -20,7 +20,6 @@ from chia.rpc.harvester_rpc_client import HarvesterRpcClient
 from chia.server.outbound_message import NodeType, make_msg
 from chia.server.start_service import Service
 from chia.simulator.block_tools import BlockTools
-from chia.simulator.time_out_assert import time_out_assert
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.peer_info import UnresolvedPeerInfo
 from chia.util.config import load_config
@@ -28,6 +27,7 @@ from chia.util.hash import std_hash
 from chia.util.ints import uint8, uint32, uint64
 from chia.util.keychain import generate_mnemonic
 from tests.conftest import HarvesterFarmerEnvironment
+from tests.util.time_out_assert import time_out_assert
 
 
 def farmer_is_started(farmer: Farmer) -> bool:
