+++ tests/plot_sync/test_plot_sync.py
@@ -24,7 +24,6 @@ from chia.protocols.harvester_protocol import Plot
 from chia.protocols.protocol_message_types import ProtocolMessageTypes
 from chia.server.start_service import Service
 from chia.simulator.block_tools import BlockTools
-from chia.simulator.time_out_assert import time_out_assert
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.util.config import create_default_chia_config, lock_and_load_config, save_config
 from chia.util.ints import uint8, uint32, uint64
@@ -32,6 +31,7 @@ from chia.util.streamable import _T_Streamable
 from tests.plot_sync.util import start_harvester_service
 from tests.plotting.test_plot_manager import Directory, MockPlotInfo
 from tests.plotting.util import get_test_plots
+from tests.util.time_out_assert import time_out_assert
 
 
 def synced(sender: Sender, receiver: Receiver, previous_last_sync_id: int) -> bool:
