+++ tests/plot_sync/util.py
@@ -12,10 +12,10 @@ from chia.plot_sync.sender import Sender
 from chia.protocols.harvester_protocol import PlotSyncIdentifier
 from chia.server.outbound_message import Message, NodeType
 from chia.server.start_service import Service
-from chia.simulator.time_out_assert import time_out_assert
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.peer_info import PeerInfo, UnresolvedPeerInfo
 from chia.util.ints import uint16, uint64
+from tests.util.time_out_assert import time_out_assert
 
 
 @dataclass
