+++ tests/simulation/test_simulation.py
@@ -22,7 +22,6 @@ from chia.simulator.keyring import TempKeyring
 from chia.simulator.setup_nodes import FullSystem, SimulatorsAndWallets
 from chia.simulator.setup_services import setup_full_node
 from chia.simulator.simulator_protocol import FarmNewBlockProtocol, GetAllCoinsProtocol, ReorgProtocol
-from chia.simulator.time_out_assert import time_out_assert
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.peer_info import PeerInfo
 from chia.util.ints import uint8, uint16, uint32, uint64
@@ -30,6 +29,7 @@ from chia.util.ws_message import create_payload
 from chia.wallet.util.tx_config import DEFAULT_TX_CONFIG
 from chia.wallet.wallet_node import WalletNode
 from tests.core.node_height import node_height_at_least
+from tests.util.time_out_assert import time_out_assert
 
 chiapos_version = pkg_resources.get_distribution("chiapos").version
