+++ tests/wallet/cat_wallet/test_trades.py
@@ -10,7 +10,6 @@ from chia.consensus.cost_calculator import NPCResult
 from chia.consensus.default_constants import DEFAULT_CONSTANTS
 from chia.full_node.bundle_tools import simple_solution_generator
 from chia.full_node.mempool_check_conditions import get_name_puzzle_conditions
-from chia.simulator.time_out_assert import time_out_assert
 from chia.types.blockchain_format.program import INFINITE_COST, Program
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.util.ints import uint32, uint64
@@ -27,6 +26,7 @@ from chia.wallet.vc_wallet.cr_cat_drivers import ProofsChecker
 from chia.wallet.vc_wallet.cr_cat_wallet import CRCATWallet
 from chia.wallet.vc_wallet.vc_store import VCProofs
 from tests.conftest import SOFTFORK_HEIGHTS, ConsensusMode
+from tests.util.time_out_assert import time_out_assert
 from tests.wallet.conftest import WalletEnvironment, WalletStateTransition, WalletTestFramework
 from tests.wallet.vc_wallet.test_vc_wallet import mint_cr_cat
