+++ tests/wallet/db_wallet/test_dl_offers.py
@@ -5,7 +5,6 @@ from typing import Any, List, Tuple
 import pytest
 
 from chia.data_layer.data_layer_wallet import DataLayerWallet
-from chia.simulator.time_out_assert import time_out_assert
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.util.ints import uint64
 from chia.wallet.puzzle_drivers import Solver
@@ -15,6 +14,7 @@ from chia.wallet.trading.trade_status import TradeStatus
 from chia.wallet.util.merkle_utils import build_merkle_tree, simplify_merkle_proof
 from chia.wallet.util.tx_config import DEFAULT_TX_CONFIG
 from tests.conftest import ConsensusMode
+from tests.util.time_out_assert import time_out_assert
 
 
 async def is_singleton_confirmed_and_root(dl_wallet: DataLayerWallet, lid: bytes32, root: bytes32) -> bool:
