+++ tests/wallet/test_notifications.py
@@ -10,13 +10,13 @@ import pytest
 from chia.consensus.block_rewards import calculate_base_farmer_reward, calculate_pool_reward
 from chia.simulator.full_node_simulator import FullNodeSimulator
 from chia.simulator.simulator_protocol import FarmNewBlockProtocol
-from chia.simulator.time_out_assert import time_out_assert, time_out_assert_not_none
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.peer_info import PeerInfo
 from chia.util.db_wrapper import DBWrapper2
 from chia.util.ints import uint32, uint64
 from chia.wallet.notification_store import NotificationStore
 from chia.wallet.util.tx_config import DEFAULT_TX_CONFIG
+from tests.util.time_out_assert import time_out_assert, time_out_assert_not_none
 
 
 # For testing backwards compatibility with a DB change to add height
