+++ tests/wallet/test_wallet_node.py
@@ -17,7 +17,6 @@ from chia.protocols.wallet_protocol import CoinState
 from chia.server.outbound_message import Message, make_msg
 from chia.simulator.block_tools import test_constants
 from chia.simulator.setup_nodes import SimulatorsAndWallets
-from chia.simulator.time_out_assert import time_out_assert
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.full_block import FullBlock
 from chia.types.mempool_inclusion_status import MempoolInclusionStatus
@@ -31,6 +30,7 @@ from chia.wallet.util.tx_config import DEFAULT_TX_CONFIG
 from chia.wallet.wallet_node import Balance, WalletNode
 from tests.conftest import ConsensusMode
 from tests.util.misc import CoinGenerator
+from tests.util.time_out_assert import time_out_assert
 
 
 @pytest.mark.anyio
